use only UserID

+++ internal/services/audit/entry.go
@@ -13,7 +13,7 @@ func NewEntryFromAuditEvent(event Event) Entry {
 	return Entry{
 		TimeTS:      event.GetTime().Unix(),
 		Action:      event.GetName(),
-		UserID:      strconv.FormatInt(event.GetUserID(), 10),
+		UserID:      strconv.FormatInt(int64(event.GetUserID()), 10),
 		OriginalURL: string(event.GetOriginalURL()),
 	}
 }
